writer-env/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
psycopg2==2.7.7
pytz==2018.9
PyYAML==5.1
